Improve request matching based on version parameter
Methods annotated with @VersionedResource are incorrectly selected when
version is expressed as range, i.e. version=~1.

The remedy for that is to gather information about all available versions
and then use it for request matching.

Although this fixes the algorithm, two tests are still failing. When
client sends Accept header with range expression, it receives in
Content-Type the same value, instead of specific version number. This is
because Spring Framework ignores media-type parameters in content
negotiation.

@see https://jira.spring.io/browse/SPR-10903